Split into containers and components
