+++ gatsby-theme-secorto/src/@scot3004/gatsby-theme-portfolio-core/components/post.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import PortfolioPost from '../../../components/portfolio-post'
+import PortfolioPost from '../../../containers/portfolio-post'
 
 export default ({ data }) => (
   <PortfolioPost data={ data } />
