+++ gatsby-theme-secorto/src/assets/gatsby-icon.png
