+++ gatsby-theme-secorto/src/assets/moon.png
