+++ gatsby-theme-secorto/src/assets/sun.png
