+++ gatsby-theme-secorto/src/components/Bio/bio-content.jsx
