+++ gatsby-theme-secorto/src/components/Post/post-hero.jsx
