+++ gatsby-theme-secorto/src/components/Post/post-title.jsx
