--- gatsby-theme-secorto/src/components/brand.jsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default ({title}) => (
-  <span>MyWebSite</span>
-)
