+++ gatsby-theme-secorto/src/components/ui/gallery.jsx
