+++ gatsby-theme-secorto/src/gatsby-plugin-theme-ui/components.js
@@ -4,7 +4,7 @@ import { jsx } from "theme-ui"
 import { preToCodeBlock } from "mdx-utils"
 import PrismCodeBlock from "@theme-ui/prism"
 
-import headings from "../components/headings"
+import headings from "../components/ui/headings"
 
 const CodeBlock = preProps => {
   const props = preToCodeBlock(preProps)
