+++ gatsby-theme-secorto/src/gatsby-theme-blog-core/components/post.jsx
@@ -1,5 +1,5 @@
 import React from "react"
-import Post from "../../components/blog-post"
+import Post from "../../containers/blog-post"
 
 export default ({ location, data }) => {
   const { blogPost, previous, next } = data
