+++ gatsby-theme-secorto/src/hooks/configOptions.js
@@ -4,7 +4,6 @@ const useBlogThemeConfig = () => {
   const data = useStaticQuery(graphql`
     query {
       blogThemeConfig(id: { eq: "gatsby-theme-blog-config" }) {
-        disableThemeUiStyling
         webfontURL
       }
     }
