+++ website/src/@scot3004/gatsby-theme-secorto/components/Bio/bio-content.jsx
